Delete identity getter with HttpUrlConnection and refactor net package
