+++ src/main/kotlin/data/net/api/RetrofitBuilder.kt
@@ -1,4 +1,4 @@
-package data.api
+package data.net.api
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
